Find and link SeriousProton
Instead of adding all of SeriousProton's files to our own object list, we now try to find SeriousProton on our system automatically. Failing that, we will build SeriousProton from scratch.
Currently the location for building SeriousProton from scratch is my own fork which has CMake support. This will have to be changed if the CMake support for SeriousProton is merged to daid's repository.