Fix Serious Proton find script
It can actually find Serious Proton now. I'm not very happy with the way it has to do the final check though, with the string finding. According to the documentation, it should mark -NOTFOUND as false and absence of aforementioned as true, but somehow that's not happening. So I implemented that myself.