Fix installation of SFML libraries
SFML_ROOT may not be set. It's completely optional, just intended as one of the hints for finding the SFML installation. In particular, it may not be set if SFML was found in one of the standard installation directories or if it was built from source. Instead, let's just use the SFML_LIBRARIES that's filled by both the find script and the custom build from source.